Remove unnecessary data from abstracted header
The abstrac layers don't need that information so extracting it and saving
it is a waste of time and space.
The deserializer should at that point have already used all that information
so it doesn't need it.
When serializing a packet we should recalculate length, extension
offsets and checksums anew anyway because that information is most
likely wrong at that point.

Saving version is useless too: If it's not `1` we can't parse the packet
correctly and we can't serialize it either.